refactor: CardList, Profile 컴포넌트분리
