+++ src/components/CardListItem.css
@@ -3,22 +3,22 @@ a {
 }
 
 .CardLinkList {
-  margin: 2.5rem auto;
+  margin: 40px auto;
   padding: 0;
   display: grid;
-  grid-template-columns: repeat(3, 21.25rem);
+  grid-template-columns: repeat(3, 340px);
   place-items: center;
   justify-content: center;
-  grid-gap: 1.5625rem 1.25rem;
+  grid-gap: 25px 20px;
   list-style: none;
 }
 
 .CardListItem {
-  width: 21.25rem;
-  height: 20.875rem;
-  box-shadow: 0rem 0.3125rem 1.5625rem 0rem rgba(0, 0, 0, 0.08);
-  border-radius: 0.9375rem;
-  gap: 1.25rem 1.5625rem;
+  width: 340px;
+  height: 334px;
+  box-shadow: 0px 5px 25px 0px rgba(0, 0, 0, 0.08);
+  border-radius: 15px;
+  gap: 20px 25px;
   overflow: hidden;
 }
 
@@ -33,8 +33,8 @@ a {
 }
 
 .CardListItem__previewImg {
-  width: 21.25rem;
-  height: 12.5rem;
+  width: 340px;
+  height: 200px;
   flex-shrink: 0;
 }
 
@@ -45,16 +45,16 @@ a {
 
 .CardListItem__content {
   display: flex;
-  padding: 0.9375rem 1.25rem;
+  padding: 15px 20px;
   flex-direction: column;
   align-items: flex-start;
-  gap: 0.625rem;
+  gap: 10px;
 }
 
 .CardListItem__postTime {
   color: #666;
   font-family: Pretendard;
-  font-size: 0.8125rem;
+  font-size: 13px;
   font-style: normal;
   font-weight: 400;
   line-height: normal;
@@ -68,33 +68,33 @@ a {
   white-space: nowrap;
 
   font-family: Pretendard;
-  font-size: 1rem;
+  font-size: 16px;
   font-style: normal;
   font-weight: 400;
-  line-height: 1.5rem;
+  line-height: 24px;
   text-decoration: none;
 }
 
 .CardListItem__postDate {
   color: #333;
   font-family: Pretendard;
-  font-size: 0.875rem;
+  font-size: 14px;
   font-style: normal;
   font-weight: 400;
   line-height: normal;
 }
 
-@media (max-width: 70.25rem) {
+@media (max-width: 1124px) {
   .CardLinkList {
-    grid-template-columns: repeat(2, 21.25rem);
+    grid-template-columns: repeat(2, 340px);
   }
 }
 
-@media (max-width: 47.9375rem) {
+@media (max-width: 767px) {
   .CardLinkList {
-    grid-template-columns: repeat(1, 20.3125rem);
+    grid-template-columns: repeat(1, 325px);
   }
   .CardListItem {
-    height: 20.4375rem;
+    height: 327px;
   }
 }
