+++ src/components/Footer.css
@@ -26,7 +26,7 @@ footer {
   color: #676767;
   text-align: center;
   font-family: Acme;
-  font-size: 1rem;
+  font-size: 16px;
   font-style: normal;
   font-weight: 400;
   line-height: normal;
@@ -40,7 +40,7 @@ footer {
 .footer__supportLink > * {
   color: #cfcfcf;
   font-family: Acme;
-  font-size: 1rem;
+  font-size: 16px;
   font-style: normal;
   font-weight: 400;
   line-height: normal;
