+++ src/components/Header.css
@@ -10,19 +10,19 @@ header {
 }
 
 .Header__profileArea__profileImg {
-  width: 3.75rem;
-  height: 3.75rem;
-  padding: 1rem 1.0625rem 1rem 0.75rem;
+  width: 60px;
+  height: 60px;
+  padding: 16px 17px 16px 12px;
 }
 
 .Header__profileArea__nickname {
   color: var(--text-color-light-mode, #000);
   font-family: Pretendard;
-  font-size: 1rem;
+  font-size: 16px;
   font-style: normal;
   font-weight: 400;
-  line-height: 1.5rem;
-  padding-bottom: 1.25rem;
+  line-height: 24px;
+  padding-bottom: 20px;
 }
 
 .Header__profileArea__folderName {
