+++ src/components/Nav.css
@@ -1,6 +1,6 @@
 .nav {
   background-color: var(--Linkbrary-bg, #f0f6ff);
-  padding: 1.25rem 12.5rem;
+  padding: 20px 200px;
 }
 
 .nav ul {
@@ -15,17 +15,17 @@
 }
 
 .nav__logo {
-  width: 12.0625rem;
-  height: 1.5rem;
+  width: 193px;
+  height: 24px;
 }
 
 .nav__signinBtn {
   display: flex;
-  width: 8rem;
-  padding: 1rem 1.25rem;
+  width: 128px;
+  padding: 16px 20px;
   justify-content: center;
   align-items: center;
-  gap: 0.625rem;
+  gap: 10px;
   border-radius: 8px;
   background: var(
     --gra-purpleblue-to-skyblue,
@@ -34,7 +34,7 @@
   border-style: none;
   color: var(--Grey-Light, #f5f5f5);
   font-family: Pretendard;
-  font-size: 1.125rem;
+  font-size: 18px;
   font-style: normal;
   font-weight: 600;
   line-height: normal;
@@ -43,27 +43,27 @@
 .nav__profile {
   display: flex;
   align-items: center;
-  gap: 0.375rem;
+  gap: 6px;
 }
 
 .nav__ProfileImg {
-  width: 1.75rem;
-  height: 1.75rem;
+  width: 28px;
+  height: 28px;
 }
 
 /* 태블릿 사이즈 */
-@media (max-width: 74.9375rem) {
+@media (max-width: 1199px) {
   .nav {
-    padding: 1.25rem 2rem;
+    padding: 20px 32px;
   }
 
   .nav ul {
     margin: auto;
-    max-width: 50rem;
+    max-width: 800px;
   }
 }
 
-@media (max-width: 47.9375rem) {
+@media (max-width: 767px) {
   .nav__ProfileEmail {
     display: none;
   }
