+++ src/components/Nav.jsx
@@ -1,12 +1,8 @@
 import logoImg from "../assets/logo.svg";
-import profileImg from "../assets/profile img.svg";
 import "./Nav.css";
-import { getUserData } from "../api/Api";
-import useData from "../hooks/useData";
+import Profile from "./Profile";
 
 function Nav() {
-  const userData = useData(getUserData);
-
   return (
     <nav className="nav">
       <ul>
@@ -16,20 +12,7 @@ function Nav() {
           </a>
         </li>
         <li>
-          {userData ? (
-            <div className="nav__profile">
-              <img
-                className="nav__ProfileImg"
-                src={profileImg}
-                alt="Profile Info"
-              />
-              <span className="nav__ProfileEmail">{userData.email}</span>
-            </div>
-          ) : (
-            <a href="/signin.html" className="nav__signinBtn">
-              로그인
-            </a>
-          )}
+          <Profile />
         </li>
       </ul>
     </nav>
