+++ src/index.jsx
@@ -1,7 +1,5 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import { BrowserRouter, Routes, Route } from "react-router-dom";
-import Folder from "./pages/Folder";
 import "./index.css";
 import App from "./App";
